shared/idmap: Split idrange
